+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `/api/v1/accounts/lookup` Mastodon API endpoint.
+- Implemented activity delivery queue.
 
 ### Changed